+++ docs/changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.2.0] - Unreleased
+
+### Added
+
+- **RustAdapter**: Introduce `aria2` option to support downloading segements via [aria2](https://aria2.github.io/).
+
 ## [1.1.0] - 2026-04-13
 
 ### Added